refactor(service): move matchLottos to service layer
1. Move matchLottos from domain to service layer
- Only handles tha matching logic, not domain data itself
- Responsibilities are now clearly: domain for data, service for logic

2. Fix duplicated numbers in Test
- Lotto enforces unique numbers

3. ETC
- SRP
- Refactored to match proper MVC pattern
